back_end/requirements.txt
asgiref==3.3.4
Django==2.2
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
